pyproject.toml
[tool.poetry]
name = "yield-farmer"
version = "0.1.0"
description = "Automate vault yield claiming and iBGT vault staking"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
web3 = "^7.8.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
logging = "^0.4.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
